Fixed a bug with platform spawning. Added basic camera following
The platform spawning did not have an upper limit. Meaning that it was constantly spawning platforms and then removing them from the world as the queue was getting full. So I just changed the conditions on when a platform spawns.

The Camera needs to be smoothed out and also adjusted so that it is impossible for the ball to fall outside the viewport.