Add swiping mechaninc. Destroy old physics bodies. Adjust player spawning on x axis and platform rotation
Swiping in different directions will now change the colour of the player. This is replaced the original idea of having buttons at the bottom of the screen.

The player can now spawn randomly on the x-axis to give the game some more variation and the possible angle of the platforms has become slightly greater in possibilities.